+++ demo/meeting/src/components/Card/Styled.tsx
@@ -1,0 +1,27 @@
+// Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+import styled from 'styled-components';
+
+export const SmallText = styled.small`
+  color: rgba(0, 0, 0, 0.5);
+`;
+
+export const StyledCard = styled.div`
+  .body {
+    margin-bottom: 1rem;
+  }
+
+  .header {
+    font-size: 1.5rem;
+  }
+
+  .title {
+    font-size: 1.25rem;
+    margin-top: 1rem;
+  }
+
+  .description {
+    margin: 1rem 0 1rem 0;
+  }
+`;